+++ client/pages/LoginPage.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 export default function LoginPage() {
     return (
         <div>Login Page</div>
